requirements.txt
validators
python-whois
dnspython
requests